+++ test/cljs/cxengage_javascript_sdk/modules/reporting_test.cljs
@@ -4,7 +4,8 @@
             [cljs.core.async :as a]
             [cxengage-javascript-sdk.internal-utils :as iu]
             [cxengage-javascript-sdk.domain.rest-requests :as rest]
-            [cxengage-javascript-sdk.interop-helpers :as ih]
+            [cljs-sdk-utils.interop-helpers :as ih]
+            [cljs-sdk-utils.api :as api]
             [cxengage-javascript-sdk.state :as st]
             [cxengage-javascript-sdk.modules.reporting :as rep]
             [cljs-uuid-utils.core :as uuid]
@@ -25,16 +26,16 @@
   (testing "stat query function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-stat-query-response [:api-response :results])]
                  (a/>! resp-chan successful-stat-query-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_] resp-chan))
+                 (set! api/api-request (fn [_] resp-chan))
                  (p/subscribe "cxengage/reporting/get-stat-query-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/stat-query {:statistic "resources-logged-in-count"}))))))
 
@@ -53,18 +54,18 @@
   (testing "get tenant capacity function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-capacity-response [:api-response :results])]
                  (a/>! resp-chan successful-capacity-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [request]
+                 (set! api/api-request (fn [request]
                                           (is (= (get request :url) tenant-capacity-url))
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-capacity-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-capacity))))))
 
@@ -72,18 +73,18 @@
   (testing "get resource capacity function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-capacity-response [:api-response :results])]
                  (a/>! resp-chan successful-capacity-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [request]
+                 (set! api/api-request (fn [request]
                                           (is (= (get request :url) resource-capacity-url))
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-capacity-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-capacity {:resource-id "3e5890f1-0fef-46e3-b59f-3271e3d83646"}))))))
 
@@ -101,17 +102,17 @@
   (testing "get available stats function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-available-response [:api-response])]
                  (a/>! resp-chan successful-available-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_]
+                 (set! api/api-request (fn [_]
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-available-stats-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-available-stats))))))
 
@@ -129,17 +130,17 @@
   (testing "get interaction function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-interaction-response [:api-response])]
                  (a/>! resp-chan successful-interaction-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_]
+                 (set! api/api-request (fn [_]
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-interaction-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-interaction {:interaction-id "2937ac8b-380d-472b-9b9e-599097ee8c0d"}))))))
 
@@ -162,19 +163,19 @@
   (testing "get contact interaction history function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-contact-history-response
                                                       [:api-response])]
                  (a/>! resp-chan successful-contact-history-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [request]
+                 (set! api/api-request (fn [request]
                                           (is (= (get request :url) contact-history-url))
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-contact-interaction-history-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-contact-interaction-history
                   {:contact-id "7749c9c0-3979-11e7-b8fc-d0f69d796523"}))))))
@@ -183,18 +184,18 @@
   (testing "get paged contact interaction history function success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-contact-history-response [:api-response])]
                  (a/>! resp-chan successful-contact-history-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [request]
+                 (set! api/api-request (fn [request]
                                           (is (= (get request :url) paged-history-url))
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/get-contact-interaction-history-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (done)))
                  (rep/get-contact-interaction-history {:contact-id "7749c9c0-3979-11e7-b8fc-d0f69d796523" :page 5}))))))
 
@@ -221,17 +222,17 @@
   (testing "add statistic subscription - batch success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)
                      pubsub-expected-response (get-in successful-batch-response [:api-response :results])]
                  (a/>! resp-chan successful-batch-response)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_]
+                 (set! api/api-request (fn [_]
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/batch-response"
                               (fn [error topic response]
                                 (is (= pubsub-expected-response (ih/kebabify response)))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (reset! stat-subs)
                                 (done)))
                  (rep/add-stat-subscription {:statistic "queue-length"}))))))
@@ -240,19 +241,19 @@
   (testing "add statistic subscription - subscription success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)]
                  (a/>! resp-chan successful-stat-sub-response)
                  (set! rep/stat-subscriptions stat-subs)
                  (set! uuid/make-random-uuid #(str "c82d912c-2034-4b9e-a92a-f175870f5d8b"))
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_]
+                 (set! api/api-request (fn [_]
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/stat-subscription-added"
                               (fn [error topic response]
                                 (is (= successful-stat-sub-response (ih/kebabify response)))
                                 (is (= @stat-subs successful-stat-subs-update))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (reset! stat-subs)
                                 (done)))
                  (rep/add-stat-subscription {:statistic "queue-length"}))))))
@@ -268,17 +269,17 @@
   (testing "add statistic subscription - subscription success"
     (async done
            (reset! p/sdk-subscriptions {})
-           (go (let [old rest/api-request
+           (go (let [old api/api-request
                      resp-chan (a/promise-chan)]
                  (a/>! resp-chan successful-stat-sub-response)
                  (set! rep/stat-subscriptions stat-subs)
                  (reset! st/sdk-state test-state)
-                 (set! rest/api-request (fn [_]
+                 (set! api/api-request (fn [_]
                                           resp-chan))
                  (p/subscribe "cxengage/reporting/stat-subscription-removed"
                               (fn [error topic response]
                                 (is (= @stat-subs successful-stat-removal))
-                                (set! rest/api-request old)
+                                (set! api/api-request old)
                                 (reset! stat-subs)
                                 (done)))
                  (rep/remove-stat-subscription {:stat-id "c82d912c-2034-4b9e-a92a-f175870f5d8b"}))))))